requirements.txt
evdev>=1.6
pyserial>=3.5
RPi.GPIO>=0.7
luma.core>=2.4
luma.oled>=3.8

